수강 삭제 방식 변경

+++ src/main/java/com/docker/backend/service/enrollment/EnrollmentService.java
@@ -45,7 +45,10 @@ public class EnrollmentService {
     public void cancelEnroll(Student student, Long courseId) {
         Enrollment enrollment = enrollmentRepository.findByStudentAndCourseId(student, courseId)
                 .orElseThrow(() -> new EntityNotFoundException("Enrollment not found"));
-//        enrollment.setStatus(Status.WITHDRAWN);
+
+        if (enrollment.getStatus() == Status.COMPLETED) {
+            return;
+        }
         enrollmentRepository.delete(enrollment);
     }
